BuyTheThingsImSelling/src/main/java/com/fdmgroup/buythethingsisell/jdbc/userfunctions/MD5Hasher.java
package com.fdmgroup.buythethingsisell.jdbc.userfunctions;

import java.security.MessageDigest;
import java.security.NoSuchAlgorithmException;

import org.apache.log4j.Logger;

public class MD5Hasher {
	
	final static Logger logger = Logger.getLogger(MD5Hasher.class);

	public static String getMD5Hashed(String password) {
		MessageDigest md = null;
		try {
			md = MessageDigest.getInstance("MD5");
		} catch (NoSuchAlgorithmException e) {
			logger.fatal("No algorithm for MD5 found");
		}
		md.update(password.getBytes());
		byte byteData[] = md.digest();
		StringBuffer sb = new StringBuffer();
		for (int i = 0; i < byteData.length; i++) {
			sb.append(Integer.toString((byteData[i] & 0xff) + 0x100, 16).substring(1));
		}
		return sb.toString();
	}
}